updated MultiTouchBasic sample for latest glNext
