+++ samples/MultiTouchBasic/src/MultiTouchBasicApp.cpp
@@ -1,0 +1,137 @@
+#include "cinder/app/App.h"
+#include "cinder/app/RendererGl.h"
+#include "cinder/gl/gl.h"
+#include "cinder/System.h"
+#include "cinder/Rand.h"
+#include "cinder/Log.h"
+
+#include <vector>
+#include <map>
+#include <list>
+
+using namespace ci;
+using namespace ci::app;
+using namespace std;
+
+struct TouchPoint {
+	TouchPoint() {}
+	TouchPoint( const vec2 &initialPt, const Color &color ) : mColor( color ), mTimeOfDeath( -1.0 ) 
+	{
+		mLine.push_back( initialPt ); 
+	}
+	
+	void addPoint( const vec2 &pt ) { mLine.push_back( pt ); }
+	
+	void draw() const
+	{
+		if( mTimeOfDeath > 0 ) // are we dying? then fade out
+			gl::color( ColorA( mColor, ( mTimeOfDeath - getElapsedSeconds() ) / 2.0f ) );
+		else
+			gl::color( mColor );
+
+		gl::draw( mLine );
+	}
+	
+	void startDying() { mTimeOfDeath = getElapsedSeconds() + 2.0f; } // two seconds til dead
+	
+	bool isDead() const { return getElapsedSeconds() > mTimeOfDeath; }
+	
+	PolyLine2f		mLine;
+	Color			mColor;
+	float			mTimeOfDeath;
+};
+
+class MultiTouchApp : public App {
+ public:
+	void	mouseDown( MouseEvent event ) override;
+	void	mouseDrag( MouseEvent event ) override;
+
+	void	touchesBegan( TouchEvent event ) override;
+	void	touchesMoved( TouchEvent event ) override;
+	void	touchesEnded( TouchEvent event ) override;
+
+	void	setup() override;
+	void	draw() override;
+
+  private:
+	map<uint32_t,TouchPoint>	mActivePoints;
+	list<TouchPoint>			mDyingPoints;
+};
+
+void prepareSettings( MultiTouchApp::Settings *settings )
+{
+	// By default, multi-touch is disabled on desktop and enabled on mobile platforms.
+	// You enable multi-touch from the SettingsFn that fires before the app is constructed.
+	settings->setMultiTouchEnabled( true );
+
+	// On mobile, if you disable multitouch then touch events will arrive via mouseDown(), mouseDrag(), etc.
+//	settings->setMultiTouchEnabled( false );
+}
+
+void MultiTouchApp::setup()
+{
+	CI_LOG_I( "MT: " << System::hasMultiTouch() << " Max points: " << System::getMaxMultiTouchPoints() );
+}
+
+void MultiTouchApp::touchesBegan( TouchEvent event )
+{
+	console() << "began: " << event << std::endl;
+
+	for( const auto &touch : event.getTouches() ) {
+		Color newColor( CM_HSV, Rand::randFloat(), 1, 1 );
+		mActivePoints.insert( make_pair( touch.getId(), TouchPoint( touch.getPos(), newColor ) ) );
+	}
+}
+
+void MultiTouchApp::touchesMoved( TouchEvent event )
+{
+	//console() << "Moved: " << event << std::endl;
+	for( const auto &touch : event.getTouches() ) {
+		mActivePoints[touch.getId()].addPoint( touch.getPos() );
+	}
+}
+
+void MultiTouchApp::touchesEnded( TouchEvent event )
+{
+	console() << "ended: " << event << std::endl;
+	for( const auto &touch : event.getTouches() ) {
+		mActivePoints[touch.getId()].startDying();
+		mDyingPoints.push_back( mActivePoints[touch.getId()] );
+		mActivePoints.erase( touch.getId() );
+	}
+}
+
+void MultiTouchApp::mouseDown( MouseEvent event )
+{
+	console() << "Mouse down: " << event.isRight() << " & " << event.isControlDown() << std::endl;
+}
+
+void MultiTouchApp::mouseDrag( MouseEvent event )
+{
+	console() << "Mouse drag pos: " << event.getPos() << std::endl;
+}
+
+void MultiTouchApp::draw()
+{
+	gl::enableAlphaBlending();
+	gl::clear( Color( 0.1f, 0.1f, 0.1f ) );
+
+	for( const auto &activePoint : mActivePoints ) {
+		activePoint.second.draw();
+	}
+
+	for( auto dyingIt = mDyingPoints.begin(); dyingIt != mDyingPoints.end(); ) {
+		dyingIt->draw();
+		if( dyingIt->isDead() )
+			dyingIt = mDyingPoints.erase( dyingIt );
+		else
+			++dyingIt;
+	}
+	
+	// draw yellow circles at the active touch points
+	gl::color( Color( 1, 1, 0 ) );
+	for( const auto &touch : getActiveTouches() )
+		gl::drawStrokedCircle( touch.getPos(), 20 );
+}
+
+CINDER_APP( MultiTouchApp, RendererGl, prepareSettings )
