+++ samples/MultiTouchBasic/xcode_ios/MultiTouchBasic-Info.plist
