+++ samples/MultiTouchBasic/xcode_ios/MultiTouchBasic_Prefix.pch
@@ -1,0 +1,10 @@
+//
+// Prefix header for all source files of the 'iPhoneTestGl' target in the 'iPhoneTestGl' project
+//
+
+/*
+#ifdef __OBJC__
+#import <Foundation/Foundation.h>
+#import <UIKit/UIKit.h>
+#endif
+*/
